fix: add IndexHNSWOptimized.cpp to faiss build sources
The undefined symbol error for IndexHNSWSQOptimized occurs because
IndexHNSWOptimized.cpp was not included in the faiss build.

This file contains the implementation of IndexHNSWSQOptimized class
which is used by CANDY algorithms but wasn't being compiled into
the faiss library.

Adding it to FAISS_SRC in faiss/CMakeLists.txt resolves the
undefined symbol: _ZN5faiss20IndexHNSWSQOptimizedC1E... error.